Extract content ops markdown renderers
